Split out NamedActor and friends into a new module
